agrega ruta moleculares
