+++ src/app/pages/productos-y-servicios/insumos/insumos.component.ts
@@ -58,20 +58,20 @@ export class InsumosComponent implements OnInit {
   // salsasChillOut: any[] = [];
   
 
-  //Purés y concentrados
-  chillOutFruitConcentrados: any[] = [];
-  chillOutJamConcentrados: any[] = [];
-  pureToraniRealSmoothieMix: any[] = [];
-  pureToraniPureeBlend: any[] = [];
-  concentradoTeaZone: any[] = [];
+  // //Purés y concentrados
+  // chillOutFruitConcentrados: any[] = [];
+  // chillOutJamConcentrados: any[] = [];
+  // pureToraniRealSmoothieMix: any[] = [];
+  // pureToraniPureeBlend: any[] = [];
+  // concentradoTeaZone: any[] = [];
 
-  //Moleculares
-  chillOutPearls: any[] = [];
-  teaZonePerlas: any[] = [];
-  chillOutJelly: any[] = [];
-  teaZoneJelly: any[] = [];
-  chillOutTapioca: any[] = [];
-  teaZoneTapioca: any[] = [];
+  // //Moleculares
+  // chillOutPearls: any[] = [];
+  // teaZonePerlas: any[] = [];
+  // chillOutJelly: any[] = [];
+  // teaZoneJelly: any[] = [];
+  // chillOutTapioca: any[] = [];
+  // teaZoneTapioca: any[] = [];
 
   //Tes y Tisanas
   tisanasFrutalesCafeEtrusca: any[] = [];
@@ -146,71 +146,71 @@ export class InsumosComponent implements OnInit {
       //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
       // );
 
-      //Filtra los concentrados ChillOut Fruit y crea un nuevo array.
-      this.chillOutFruitConcentrados = this.insumosArray.filter(insumo =>
-        this.contienePalabrasChillOutFruit(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // //Filtra los concentrados ChillOut Fruit y crea un nuevo array.
+      // this.chillOutFruitConcentrados = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasChillOutFruit(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      //Filtra los concentrados ChillOut Jam y crea un nuevo array.
-      this.chillOutJamConcentrados = this.insumosArray.filter(insumo =>
-        this.contienePalabrasChillOutJam(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // //Filtra los concentrados ChillOut Jam y crea un nuevo array.
+      // this.chillOutJamConcentrados = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasChillOutJam(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      //Filtra los concentrados Torani Real Smoothie Blend y crea un nuevo array.
-      this.pureToraniRealSmoothieMix = this.insumosArray.filter(insumo =>
-        this.contienePalabrasToraniSmoothieConcentrado(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // //Filtra los concentrados Torani Real Smoothie Blend y crea un nuevo array.
+      // this.pureToraniRealSmoothieMix = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasToraniSmoothieConcentrado(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra los pures Torani Puree Blend y crea un nuevo array.
-      this.pureToraniPureeBlend = this.insumosArray.filter(insumo =>
-        this.contienePalabrasToraniPureeBlendConcentrado(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra los pures Torani Puree Blend y crea un nuevo array.
+      // this.pureToraniPureeBlend = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasToraniPureeBlendConcentrado(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra los concentrados TeaZone y crea un nuevo array
-      this.concentradoTeaZone = this.insumosArray.filter(insumo =>
-        this.contienePalabrasTeaZonePures(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra los concentrados TeaZone y crea un nuevo array
+      // this.concentradoTeaZone = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasTeaZonePures(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra los chillOut Pearls y crea un nuevo array
-      this.chillOutPearls = this.insumosArray.filter(insumo =>
-        this.contienePalabrasChillOutPearls(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra los chillOut Pearls y crea un nuevo array
+      // this.chillOutPearls = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasChillOutPearls(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra los TeaZone Perlas crea un nuevo array
-      this.teaZonePerlas = this.insumosArray.filter(insumo =>
-        this.contienePalabrasTeaZonePerlas(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra los TeaZone Perlas crea un nuevo array
+      // this.teaZonePerlas = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasTeaZonePerlas(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra los Jellys ChillOut y crea un nuevo array
-      this.chillOutJelly = this.insumosArray.filter(insumo =>
-        this.contienePalabrasChillOutJelly(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra los Jellys ChillOut y crea un nuevo array
+      // this.chillOutJelly = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasChillOutJelly(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra los Jellys Tea Zone y crea un nuevo array
-      this.teaZoneJelly = this.insumosArray.filter(insumo =>
-        this.contienePalabrasTeaZoneJelly(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra los Jellys Tea Zone y crea un nuevo array
+      // this.teaZoneJelly = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasTeaZoneJelly(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra las Tapiocas Chillout y crea un nuevo array
-      this.chillOutTapioca = this.insumosArray.filter(insumo =>
-        this.contienePalabrasChillOutTapioca(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra las Tapiocas Chillout y crea un nuevo array
+      // this.chillOutTapioca = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasChillOutTapioca(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
-      // Filtra las Tapiocas Chillout y crea un nuevo array
-      this.teaZoneTapioca = this.insumosArray.filter(insumo =>
-        this.contienePalabrasTeaZoneTapioca(insumo.name) &&
-        insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
-      );
+      // // Filtra las Tapiocas Chillout y crea un nuevo array
+      // this.teaZoneTapioca = this.insumosArray.filter(insumo =>
+      //   this.contienePalabrasTeaZoneTapioca(insumo.name) &&
+      //   insumo.urlArticleFirst && insumo.urlArticleFirst.trim() !== ''
+      // );
 
       // Filtra las Tisanas Café Etrusca y crea un nuevo array
       this.tisanasFrutalesCafeEtrusca = this.insumosArray.filter(insumo =>
@@ -373,188 +373,188 @@ export class InsumosComponent implements OnInit {
   //   return cumpleCondicion;
   // }
 
-  contienePalabrasChillOutFruit(nombre: string): boolean {
-    const palabrasClave = ['CHILLOUT', 'FRUIT', 'CONCENTRADO'];
+  // contienePalabrasChillOutFruit(nombre: string): boolean {
+  //   const palabrasClave = ['CHILLOUT', 'FRUIT', 'CONCENTRADO'];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['CHAMOY'];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['CHAMOY'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasChillOutJam(nombre: string): boolean {
-    const palabrasClave = ['CHILLOUT', 'JAM'];
+  // contienePalabrasChillOutJam(nombre: string): boolean {
+  //   const palabrasClave = ['CHILLOUT', 'JAM'];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['ACTIVE'];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['ACTIVE'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasToraniSmoothieConcentrado(nombre: string): boolean {
-    const palabrasClave = ['TORANI', 'REAL', 'SMOOTHIE'];
+  // contienePalabrasToraniSmoothieConcentrado(nombre: string): boolean {
+  //   const palabrasClave = ['TORANI', 'REAL', 'SMOOTHIE'];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['JARABE', 'BLEND'];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['JARABE', 'BLEND'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasToraniPureeBlendConcentrado(nombre: string): boolean {
-    const palabrasClave = ['TORANI', 'PURE', 'BLEND'];
+  // contienePalabrasToraniPureeBlendConcentrado(nombre: string): boolean {
+  //   const palabrasClave = ['TORANI', 'PURE', 'BLEND'];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['SMOOTHIE', 'JARABE'];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['SMOOTHIE', 'JARABE'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasTeaZonePures(nombre: string): boolean {
-    const palabrasClave = ['TEA', 'ZONE', 'CONCENTRADO'];
+  // contienePalabrasTeaZonePures(nombre: string): boolean {
+  //   const palabrasClave = ['TEA', 'ZONE', 'CONCENTRADO'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasChillOutPearls(nombre: string): boolean {
-    const palabrasClave = ['CHILLOUT', 'PEARL',];
+  // contienePalabrasChillOutPearls(nombre: string): boolean {
+  //   const palabrasClave = ['CHILLOUT', 'PEARL',];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['JAM', 'JARABE', 'TAPIOCA'];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['JAM', 'JARABE', 'TAPIOCA'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasTeaZonePerlas(nombre: string): boolean {
-    const palabrasClave = ['TEA', 'ZONE', 'PERLAS'];
+  // contienePalabrasTeaZonePerlas(nombre: string): boolean {
+  //   const palabrasClave = ['TEA', 'ZONE', 'PERLAS'];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['CONCENTRADO', 'TAPIOCA',];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['CONCENTRADO', 'TAPIOCA',];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasChillOutJelly(nombre: string): boolean {
-    const palabrasClave = ['CHILLOUT', 'JELLY'];
+  // contienePalabrasChillOutJelly(nombre: string): boolean {
+  //   const palabrasClave = ['CHILLOUT', 'JELLY'];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['CONCENTRADO', 'TAPIOCA', 'JARABE'];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['CONCENTRADO', 'TAPIOCA', 'JARABE'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasTeaZoneJelly(nombre: string): boolean {
-    const palabrasClave = ['TEA', 'ZONE', 'JELLY'];
+  // contienePalabrasTeaZoneJelly(nombre: string): boolean {
+  //   const palabrasClave = ['TEA', 'ZONE', 'JELLY'];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['PERLAS', 'CONCENTRADO',];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['PERLAS', 'CONCENTRADO',];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasChillOutTapioca(nombre: string): boolean {
-    const palabrasClave = ['TAPIOCA',];
+  // contienePalabrasChillOutTapioca(nombre: string): boolean {
+  //   const palabrasClave = ['TAPIOCA',];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['JELLY', 'TEA', 'ZONE'];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['JELLY', 'TEA', 'ZONE'];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
-  contienePalabrasTeaZoneTapioca(nombre: string): boolean {
-    const palabrasClave = ['TAPIOCA', 'ZONE', 'TEA',];
+  // contienePalabrasTeaZoneTapioca(nombre: string): boolean {
+  //   const palabrasClave = ['TAPIOCA', 'ZONE', 'TEA',];
 
-    // Palabras a excluir
-    const palabrasExcluidas = ['CHILOUT', 'CHILLOUT',];
+  //   // Palabras a excluir
+  //   const palabrasExcluidas = ['CHILOUT', 'CHILLOUT',];
 
-    // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
-    const nombreEnMayusculas = nombre.toUpperCase();
+  //   // Convierte el nombre a mayúsculas para hacer la coincidencia sin distinción de mayúsculas y minúsculas
+  //   const nombreEnMayusculas = nombre.toUpperCase();
 
-    // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
-    const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
-      && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
+  //   // Verifica que todas las palabras clave estén presentes y que ninguna palabra excluida esté presente
+  //   const cumpleCondicion = palabrasClave.every(palabra => nombreEnMayusculas.includes(palabra))
+  //     && !palabrasExcluidas.some(excluida => nombreEnMayusculas.includes(excluida));
 
-    //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
-    return cumpleCondicion;
-  }
+  //   //console.log(nombre, cumpleCondicion); // Verifica los resultados en la consola
+  //   return cumpleCondicion;
+  // }
 
   contienePalabrasTisanaFrutalesCafeEtrusca(nombre: string): boolean {
     const palabrasClave = ['TISANA', 'FRUTAL'];
